% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\docType{package}
\name{aggregation}
\alias{aggregation}
\title{aggregation: A package for computing the number of individuals
detected by the network.}
\description{
This package contains functions to compute the  number of individuals
detected by the network in each geogrphical region as well as the number of
individuals detected by the network moving from one region to another at
successive time instants. For an example on how to use this package please read
\link{example}.
}
