% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPostLocJointProbs.R
\name{buildPostLocJointProbs}
\alias{buildPostLocJointProbs}
\title{Builds the table with joint posterior location probabilities from simulated data.}
\usage{
buildPostLocJointProbs(
  path,
  simFileName,
  gridFileName,
  eventsFileName,
  signalFileName
)
}
\arguments{
\item{simFileName}{}

\item{gridFileName}{}

\item{eventsFileName}{}

\item{signalFileName}{}
}
\description{
Builds the table with joint posterior location probabilities from simulated data.
}
