% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisher}
\alias{fisher}
\title{Fisher's Method}
\usage{
fisher(pvalues)
}
\arguments{
\item{pvalues}{A vector of p-values (i.e. between 0 and 1) to be aggregated with Fisher's method. NAs will be filtered out.}
}
\description{
Aggregate p-values with equal weights. Equivalent to the Lancaster method with all p-values weighted at 2.
}
\examples{
fisher(c(.1, .2, .3))
}
