% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancaster.R
\name{lancaster}
\alias{lancaster}
\title{Lancaster method}
\usage{
lancaster(pvalues, weights)
}
\arguments{
\item{pvalues}{A vector of p-values (i.e. between 0 and 1). NAs will be filtered out.}

\item{weights}{A vector of weights, each associated with its respective p-value. Weights must be nonegative. NAs and negative weights will be filtered out with corresponding p-values.}
}
\description{
Weighted p-value aggregation.
}
\examples{
lancaster(c(.1, .5), c(2, 4))
}
