% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nIndividuals2.R
\name{nIndividuals2}
\alias{nIndividuals2}
\title{Computes the number of individuals moving from a region to another as
  a sum of multinomial variables.}
\usage{
nIndividuals2(categories, cellNames)
}
\arguments{
\item{categories}{a list of categories in the form \code{region-devCount}.}

\item{cellNames}{A vector with the names of the regions.}
}
\value{
The sum of the multinomial variates.
}
\description{
Computes the number of individuals moving from a region to
  another as a sum of multinomial varibales. It is an internal function that
  is called during the process of generating the Poisson multinomial
  distributed random values. This function cannot be directly called by
  users.
}
