% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNnet_Event.R
\name{rNnet_Event}
\alias{rNnet_Event}
\title{Generates random values from a Poisson multinomial distribution.}
\usage{
rNnet_Event(n, prob.dt)
}
\arguments{
\item{n}{The number of random values to generate.}

\item{prob.dt}{A data.table object with the following columns:\code{device,
cell, devCount, prob}.}
}
\value{
A matrix object with the random values generated according to a
  Poisson multinomial distribution.
}
\description{
Generates random values from a Poisson multinomial distribution.
  This function is only called internally. It cannot be called directly by
  users.
}
