% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNnetEvent.R
\name{rNnetEvent}
\alias{rNnetEvent}
\title{Generates random values according to a Poisson multinomial probability distribution.}
\usage{
rNnetEvent(n, dupFileName, regsFileName, postLocPath, prefix, times = NULL)
}
\arguments{
\item{n}{The number of values to be generated}

\item{dupFileName}{The name of the .csv file with the duplicity probability for each device}

\item{regsFileName}{The name of the .csv file defining the regions. It has two columns: \code{ tile, region}. The
first column contains the IDs of each tile in the grid while the second contains the number of a region.}

\item{postLocPath}{The path where the files with the posterior location probabilities for each devices can be found.
A file with the location probabilities should have the name \code{postLocDevice_ID.csv} where \code{ID} is replaced
with the device ID.}

\item{prefix}{TODO}

\item{times}{An optional vector with the time instants when the events were registered.}
}
\description{
Generates random values according to a Poisson multinomial probability distribution. A point estimation
derived from this distribution (mean, mode) represents an estimation of the number of individuals in a region.
}
