% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidak.R
\name{sidak}
\alias{sidak}
\title{Perform the Sidak method.}
\usage{
sidak(pvalues)
}
\arguments{
\item{pvalues}{A vector of p-values to be aggregated. NAs will be filtered.}
}
\description{
The Sidak method uses the minimum p-value but corrects it for the number of p-values that are aggregated.
}
\examples{
sidak(c(.1, .2, .3))
}
