% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPostLocJointProbs.R
\name{buildPostLocJointProbs}
\alias{buildPostLocJointProbs}
\title{Builds the table with joint posterior location probabilities from
 simulated data.}
\usage{
buildPostLocJointProbs(
  path,
  simFileName,
  gridFileName,
  eventsFileName,
  signalFileName
)
}
\arguments{
\item{path}{The path of the directory where the outfiles are saved.}

\item{simFileName}{The name of the file with the general simulation parameters.
parameters.}

\item{gridFileName}{The name of the file with the grid parameters.}

\item{eventsFileName}{The name of the file containing the networks events.}

\item{signalFileName}{The name of the file with the signal strength/quality.}
}
\description{
This is a utility function with the purpose to build the files
 with joint posterior location probabilities for each device using the
 \code{destim} package an a series of additional data. This function was used
 to build the example data set included in this package The raw data were
 produced using the simulation software. The joint posterior location
 probabilities file names starts with \code{postLocJointProbDevice_} followed for
 the \code{deviceID} and the extension \code{.dt.csv}.
}
