% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAggr.R
\name{doAggr}
\alias{doAggr}
\title{Generates a sequence of Poisson multinomial random values for a
  sequence of time instants.}
\usage{
doAggr(ichunks, n, tiles, postLoc, dupProbs, regions)
}
\arguments{
\item{ichunks}{A partition of time instants.}

\item{n}{The number of random values to be generated.}

\item{tiles}{The list of tile indexes.}

\item{postLoc}{A list with posterior location probabilities matrices. Each
element of the list corresponds to a device.}

\item{dupProbs}{The duplicity probabilities for each device. It is a
data.table object with two columns: \code{deviceID, dupProb}.}

\item{regions}{a data.table object with two columns: \code{tile, region}. It
defines each region as a composition of multiple tiles.}
}
\value{
a data.table object with the following column: \code{time, region,
  N}. It contains \code{n} random values for each combination
  \code{time-region}. An estimation of the number of individuals in a region
  can be obtained using the mean, mode or median of these values.
}
\description{
Generates a sequence of Poisson multinomial random values for a
  sequence of time instants. This function is called internally by
  \code{rNnetEvent} function which partitions the time instants into equal
  chunks, creates a cluster of working processes and distributes the chunks
  to different workers that run in parallel. This is an internal function
  that cannot be directly called by the users.
}
