% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNnetEventOD.R
\name{rNnetEventOD}
\alias{rNnetEventOD}
\title{Generates random value according to a Poisson multinomial distribution}
\usage{
rNnetEventOD(n, dupFileName, regsFileName, postLocJointPath, prefix)
}
\description{
Generates random value according to a Poisson multinomial distribution
}
